# refactor performance specs